+++ setup-libs.ps1
@@ -55,6 +55,7 @@ if (-not (Test-Path $R2ModmanProfilePath)) {
 $r2Dlls = @(
     "BepInEx\core\BepInEx.dll"
     "BepInEx\plugins\EternalsTeam-PluginConfigurator\PluginConfigurator\PluginConfigurator.dll"
+    "BepInEx\plugins\end_4-Notiffy\Notiffy.dll"
 )
 
 foreach ($dll in $r2Dlls) {