+++ package/CHANGELOG.md
@@ -1,5 +1,13 @@
 # ActivateNocturnalOS
 
+## 1.1.0
+
+- Added random nag notifications
+  - FourDrive
+  - Smilesoft365
+  - Genuine license
+- Added a gradient to the icon :D
+
 ## 1.0.0
 
-- Initial release
+- Initial release: Bottom-right text overlay